Rename and remove java files to corresponding packages
