+++ server/src/main/java/com/lunchtime/controllers/FeedbackController.java
@@ -3,12 +3,10 @@ package com.lunchtime.controllers;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
-import java.util.Optional;
-import com.lunchtime.dto.FeedbackDto;
+
+import com.lunchtime.service.dto.FeedbackDto;
 import com.lunchtime.models.Feedback;
-import com.lunchtime.implementation.FeedbackServiceImplement;
-import com.lunchtime.models.Restaurant;
-import org.springframework.http.HttpStatus;
+import com.lunchtime.service.impl.FeedbackServiceImpl;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,9 +17,9 @@ import org.springframework.web.bind.annotation.*;
 public class FeedbackController {
 
     //TODO rename variable name to normal CamelCase
-    private final FeedbackServiceImplement feedbackServiceimplementation;
+    private final FeedbackServiceImpl feedbackServiceimplementation;
 
-    public FeedbackController(FeedbackServiceImplement feedbackServiceimplementation) {
+    public FeedbackController(FeedbackServiceImpl feedbackServiceimplementation) {
         this.feedbackServiceimplementation = feedbackServiceimplementation;
     }
